AWESOME! Lets move to members side this time, and add more tabs such as: my attendance, my fees details, 'pay fees' which should include all the payment options etc. or have an option to open camera to scan a UPI qr code at the library desk. the monthly overview should be seperate and include total hours studied.
Also, we forgot to add the fee details on each student details in the admin part. And also about fees due and a list of students (in order) of dues. It should display next to them when they last payed, and when is their next payment date.

Then, total revenue should be on the dashboard.